backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.6.1
pydantic>=2.5.0
python-jose==3.3.0
httpx>=0.24.1 